checkpoint

+++ clusterresult.h
@@ -32,7 +32,7 @@ public:
 	virtual ~cluster_result();
 	vector<cluster_info*> clusters;
 	cluster_info * new_cluster(long cluster_id);
-	int partition_id;
+	long partition_id;
 	void print();
 };
